+++ codemp/rd-vanilla/tr_bsp.cpp
@@ -383,7 +383,7 @@ static void ParseFace( dsurface_t *ds, mapVert_t *verts, msurface_t *surf, int *
 	numIndexes = LittleLong( ds->numIndexes );
 
 	// create the srfSurfaceFace_t
-	sfaceSize = ( size_t ) &((srfSurfaceFace_t *)0)->points[numPoints];
+	sfaceSize = static_cast<int>(( size_t ) &((srfSurfaceFace_t *)0)->points[numPoints]);
 	ofsIndexes = sfaceSize;
 	sfaceSize += sizeof( int ) * numIndexes;
 
@@ -1928,7 +1928,7 @@ void R_LoadEntities( lump_t *l, world_t &worldData ) {
 
 		// check for remapping of shaders for vertex lighting
 		s = "vertexremapshader";
-		if (!Q_strncmp(keyname, s, strlen(s)) ) {
+		if (!Q_strncmp(keyname, s, static_cast<int>(strlen(s))) ) {
 			s = strchr(value, ';');
 			if (!s) {
 				ri.Printf( PRINT_ALL, S_COLOR_YELLOW  "WARNING: no semi colon in vertexshaderremap '%s'\n", value );
@@ -1942,7 +1942,7 @@ void R_LoadEntities( lump_t *l, world_t &worldData ) {
 		}
 		// check for remapping of shaders
 		s = "remapshader";
-		if (!Q_strncmp(keyname, s, strlen(s)) ) {
+		if (!Q_strncmp(keyname, s, static_cast<int>(strlen(s))) ) {
 			s = strchr(value, ';');
 			if (!s) {
 				ri.Printf( PRINT_ALL, S_COLOR_YELLOW  "WARNING: no semi colon in shaderremap '%s'\n", value );
