+++ shared/sys/sys_event.cpp
@@ -75,7 +75,7 @@ sysEvent_t Sys_GetEvent( void ) {
 		char	*b;
 		int		len;
 
-		len = strlen( s ) + 1;
+		len = static_cast<int>(strlen( s ) + 1);
 		b = (char *)Z_Malloc( len,TAG_EVENT,qfalse );
 		strcpy( b, s );
 		Sys_QueEvent( 0, SE_CONSOLE, 0, 0, len, b );
